#4 Refactor into a ports-and-adapters-like architecture.
