+++ src/main.py
@@ -3,7 +3,7 @@ from os import environ
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 
-from src.application.adapters.api.routes import router
+from src.adapters.api.routes import router
 
 
 def create_app() -> FastAPI:
